requirements.txt
opencv-python
numpy
pycaw
comtypes
pillow
